Добавил код удаление категорий

+++ admin/categories/index.php
@@ -77,27 +77,35 @@ if (isset($_GET['editform'])) {
 }
 /* ........................................................................ */
 
-//Удаляем все записи, связывающие шутки с этой категорией.
-try {
-    $sql = 'DELETE FROM jokecategory WHERE categoryid = :id';
-    $s = $pdo->prepare($sql);
-    $s->bindValue(':id', $_POST['id']);
-    $s->execute();
-}
-catch (PDOException $e) {
-    $error = 'Ошибка при удалении шуток из категории.' . $e->getMessage();
-    include 'error.html.php';
+/* ...................... Удаление категорий ........................ */
+if (isset($_POST['action']) and $_POST['action'] == 'Удалить') {
+    include $_SERVER['DOCUMENT_ROOT'] . '/includes/db.inc.php';
+
+    //Удаляем все записи, связывающие шутки с этой категорией.
+    try {
+        $sql = 'DELETE FROM jokecategory WHERE categoryid = :id';
+        $s = $pdo->prepare($sql);
+        $s->bindValue(':id', $_POST['id']);
+        $s->execute();
+    }
+    catch (PDOException $e) {
+        $error = 'Ошибка при удалении шуток из категории.' . $e->getMessage();
+        include 'error.html.php';
+        exit();
+    }
+    //Удаляем категорию.
+    try {
+        $sql = 'DELETE FROM category WHERE id = :id';
+        $s = $pdo->prepare($sql);
+        $s->bindValue(':id', $_POST['id']);
+        $s->execute();
+    }
+    catch (PDOException $e) {
+        $error = 'Ошибка при удалении категории.' . $e->getMessage();
+        include 'error.html.php';
+        exit();
+    }
+    header('Location: .');
     exit();
 }
-//Удаляем категорию.
-try {
-    $sql = 'DELETE FROM category WHERE id = :id';
-    $s = $pdo->prepare($sql);
-    $s->bindValue(':id', $_POST['id']);
-    $s->execute();
-}
-catch (PDOException $e) {
-    $error = 'Ошибка при удалении категории.' . $e->getMessage();
-    include 'error.html.php';
-    exit();
-}
+/* ............................................................................. */
